requirements.txt
boto3==1.26.69
Django==4.1.6
djangorestframework==3.14.0
drf-yasg==1.21.5
